Add minimum working example thesis
Include dummy files for front matter, main content, and an appendix to
demonstrate the build results for a thesis and any other piecewise
compilation of chapters. Also include sample figures, a user-defined
metadata file, a bibliography, and a bibliography style sheet to
demonstrate how these files are incorporated into the final build.