Make creation of directories to be OS-agnostic
Previously, we ensured subdirectories of target files existed through
an explicit call to `mkdir -p` as the first rule of a recipe. This has
two problems: it is called on every invocation regardless of whether the
parent directory exists, and it fails under Windows systems because
Windows systems do not recognize the `-p` flag.

Now, parent directories are explicitly supplied as an order-only
prerequisite to each target, and a single `mkdir` function is used to
create directories only when they do not already exist.

The use of an order-only prerequisite means that the timestamp of the
directory is ignored, which avoids the pitfall where the presence of any
new file in the prerequisite directory forcefully triggers a rebuild.

One inelegant drawback of this approach is that recipes become more
specific and more difficult to refactor.