Fix make draft append settings to header-includes
Bug: Drafts do not compile with line numbers from a fresh state

The issue comes from an improper heirachy of demands. The step to build
a temporary metadata yaml file belongs as a separate dependency to the
draft recipe. Now, a _tmp/draft.yaml file represents the combined
metadata from all user supplied settings in one file, which can be
triggered to build before a draft recipe.